model-image-prediction/requirements.txt
tensorflow==2.7.0
numpy==1.21.2
protobuf<3.20.0
pillow==9.5.0
Flask==2.1.1
